move parse error handling into `parse/` package
consolidates the parse error handling and definitions into the `parse/`
package, which does not import the `api/` package.
