+++ api/flexstrings.go
@@ -6,6 +6,8 @@ package api
 
 import (
 	"gopkg.in/yaml.v3"
+
+	"github.com/gdt-dev/core/parse"
 )
 
 // FlexStrings is a struct used to parse an interface{} that can be either a
@@ -23,7 +25,7 @@ func (f *FlexStrings) Values() []string {
 // FlexStrings can be either a string or a slice of strings.
 func (f *FlexStrings) UnmarshalYAML(node *yaml.Node) error {
 	if node.Kind != yaml.ScalarNode && node.Kind != yaml.SequenceNode {
-		return ExpectedScalarOrSequenceAt(node)
+		return parse.ExpectedScalarOrSequenceAt(node)
 	}
 	var single string
 	if err := node.Decode(&single); err == nil {
@@ -35,5 +37,5 @@ func (f *FlexStrings) UnmarshalYAML(node *yaml.Node) error {
 		f.values = many
 		return nil
 	}
-	return ExpectedScalarOrSequenceAt(node)
+	return parse.ExpectedScalarOrSequenceAt(node)
 }
