+++ plugin/exec/defaults.go
@@ -5,9 +5,8 @@
 package exec
 
 import (
+	"github.com/gdt-dev/core/parse"
 	"gopkg.in/yaml.v3"
-
-	"github.com/gdt-dev/core/api"
 )
 
 type execDefaults struct{}
@@ -19,21 +18,21 @@ type Defaults struct {
 
 func (d *Defaults) UnmarshalYAML(node *yaml.Node) error {
 	if node.Kind != yaml.MappingNode {
-		return api.ExpectedMapAt(node)
+		return parse.ExpectedMapAt(node)
 	}
 	// maps/structs are stored in a top-level Node.Content field which is a
 	// concatenated slice of Node pointers in pairs of key/values.
 	for i := 0; i < len(node.Content); i += 2 {
 		keyNode := node.Content[i]
 		if keyNode.Kind != yaml.ScalarNode {
-			return api.ExpectedScalarAt(keyNode)
+			return parse.ExpectedScalarAt(keyNode)
 		}
 		key := keyNode.Value
 		valNode := node.Content[i+1]
 		switch key {
 		case "exec":
 			if valNode.Kind != yaml.MappingNode {
-				return api.ExpectedMapAt(valNode)
+				return parse.ExpectedMapAt(valNode)
 			}
 			ed := execDefaults{}
 			if err := valNode.Decode(&ed); err != nil {
