+++ scenario/from.go
@@ -9,7 +9,6 @@ import (
 
 	"gopkg.in/yaml.v3"
 
-	"github.com/gdt-dev/core/api"
 	"github.com/gdt-dev/core/parse"
 )
 
@@ -35,7 +34,7 @@ func FromBytes(
 	s := New(mods...)
 	expanded := parse.ExpandWithFixedDoubleDollar(string(contents))
 	if err := yaml.Unmarshal([]byte(expanded), s); err != nil {
-		if ep, ok := err.(*api.ParseError); ok {
+		if ep, ok := err.(*parse.Error); ok {
 			ep.Path = s.Path
 			ep.SetContents()
 			return nil, ep
